requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.1
billiard==3.6.1.0
bs4==0.0.1
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
DateTime==4.3
Django==1.11.11
django-cors-headers==3.2.0
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
Pillow==6.2.1
pycparser==2.19
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
zope.interface==4.7.1
